requirements.txt
# Core dependencies
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-docx>=0.8.11
matplotlib>=3.5.0
numpy>=1.21.0
requests>=2.28.0
lxml>=4.9.0
PyYAML>=6.0

# LLM dependencies
llama-cpp-python>=0.2.0
sentence-transformers>=2.2.2

# GUI dependencies
PySide6>=6.5.0
darkdetect>=0.8.0

# Development dependencies
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0